Add test harness fit test to graph tasks (#1301)
