+++ flash/graph/embedding/model.py
@@ -19,6 +19,7 @@ from torch import nn
 from flash.core.data.io.input import DataKeys
 from flash.core.model import Task
 from flash.graph.classification.model import GraphClassifier, POOLING_FUNCTIONS
+from flash.graph.collate import _pyg_collate
 
 
 class GraphEmbedder(Task):
@@ -41,6 +42,8 @@ class GraphEmbedder(Task):
         self.backbone = backbone
 
         self.pooling_fn = POOLING_FUNCTIONS[pooling_fn] if isinstance(pooling_fn, str) else pooling_fn
+
+        self.collate_fn = _pyg_collate
 
     def forward(self, data) -> torch.Tensor:
         x = self.backbone(data.x, data.edge_index)
