+++ tests/graph/classification/test_model.py
@@ -18,6 +18,7 @@ import torch
 
 from flash import RunningStage, Trainer
 from flash.core.data.data_module import DataModule
+from flash.core.data.io.input import DataKeys
 from flash.core.utilities.imports import _GRAPH_AVAILABLE, _GRAPH_TESTING
 from flash.graph.classification import GraphClassifier
 from flash.graph.classification.input import GraphClassificationDatasetInput
@@ -51,26 +52,11 @@ class TestGraphClassifier(TaskTester):
         assert isinstance(output, torch.Tensor)
         assert output.shape == torch.Size([1, 2])
 
-
-@pytest.mark.skipif(not _GRAPH_TESTING, reason="pytorch geometric isn't installed")
-def test_smoke():
-    """A simple test that the class can be instantiated."""
-    model = GraphClassifier(num_features=1, num_classes=1)
-    assert model is not None
-
-
-@pytest.mark.skipif(not _GRAPH_TESTING, reason="pytorch geometric isn't installed")
-def test_train(tmpdir):
-    """Tests that the model can be trained on a pytorch geometric dataset."""
-    tudataset = datasets.TUDataset(root=tmpdir, name="KKI")
-    model = GraphClassifier(num_features=tudataset.num_features, num_classes=tudataset.num_classes)
-    datamodule = DataModule(
-        GraphClassificationDatasetInput(RunningStage.TRAINING, tudataset),
-        transform=GraphClassificationInputTransform,
-        batch_size=4,
-    )
-    trainer = Trainer(default_root_dir=tmpdir, fast_dev_run=True)
-    trainer.fit(model, datamodule=datamodule)
+    @property
+    def example_train_sample(self):
+        edge_index = torch.tensor([[0, 1, 1, 2], [1, 0, 2, 1]], dtype=torch.long)
+        x = torch.tensor([[-1], [0], [1]], dtype=torch.float)
+        return {DataKeys.INPUT: Data(x=x, edge_index=edge_index), DataKeys.TARGET: 1}
 
 
 @pytest.mark.skipif(not _GRAPH_TESTING, reason="pytorch geometric isn't installed")
